requirements.txt
langgraph==0.2.70
langchain==0.3.14
langchain-google-genai==2.0.9
langchain-core==0.3.28
python-dotenv==1.0.1